pyproject.toml
[project]
name = "nibetester"
version = "0.1.0"
description = "Python RS-485 serial communication for Nibe 360P heat pump"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
]
